+++ tests/test_json.py
@@ -84,7 +84,7 @@ class User(Model):
 
     url = URLField()
     txt = TextField()
-    txt2 = TextField(multi_language=True)
+    txt2 = TextField(is_multilingual=True)
     slug = SlugField()
     phone = PhoneField()
     password = PasswordField()